move the memory functions to the library
