+++ lib.c
@@ -117,82 +117,6 @@ void mutt_nocurses_error(const char *fmt, ...)
   fputc('\n', stderr);
 }
 
-void *safe_calloc(size_t nmemb, size_t size)
-{
-  void *p = NULL;
-
-  if (!nmemb || !size)
-    return NULL;
-
-  if (((size_t) -1) / nmemb <= size)
-  {
-    mutt_error(_("Integer overflow -- can't allocate memory!"));
-    sleep(1);
-    mutt_exit(1);
-  }
-
-  if (!(p = calloc(nmemb, size)))
-  {
-    mutt_error(_("Out of memory!"));
-    sleep(1);
-    mutt_exit(1);
-  }
-  return p;
-}
-
-void *safe_malloc(size_t siz)
-{
-  void *p = NULL;
-
-  if (siz == 0)
-    return 0;
-  if ((p = malloc(siz)) == NULL)
-  {
-    mutt_error(_("Out of memory!"));
-    sleep(1);
-    mutt_exit(1);
-  }
-  return p;
-}
-
-void safe_realloc(void *ptr, size_t siz)
-{
-  void *r = NULL;
-  void **p = (void **) ptr;
-
-  if (siz == 0)
-  {
-    if (*p)
-    {
-      free(*p);
-      *p = NULL;
-    }
-    return;
-  }
-
-  r = realloc(*p, siz);
-  if (!r)
-  {
-    mutt_error(_("Out of memory!"));
-    sleep(1);
-    mutt_exit(1);
-  }
-
-  *p = r;
-}
-
-void safe_free(void *ptr)
-{
-  if (!ptr)
-    return;
-  void **p = (void **) ptr;
-  if (*p)
-  {
-    free(*p);
-    *p = 0;
-  }
-}
-
 int safe_fclose(FILE **f)
 {
   int r = 0;
