+++ po/POTFILES.in
@@ -64,6 +64,7 @@ lib/lib_base64.c
 lib/lib_date.c
 lib/lib_exit.c
 lib/lib_md5.c
+lib/lib_memory.c
 lib/lib_sha1.c
 main.c
 mbox.c
